Keep the soft keyboard visible while adding items
Now you can add an item with the enter button on the keyboard AND the
"+" button on the screen.
When pressing the "+" button on the screen, the input box would normally
loose focus causing the keyboard on a phone to disappear.
This change fixes that, see ref in code for details.