Remove the delete buttons (for now)
This is a nice feature and the animations are great, but its hard to
differentiate between clearing the ticked off items and deleting items
with button icons, so for now I think clearing the ticked items is a
more common use case, so the deleting individual items can go for now.
This will probably get re-introduced when I work out a nice way of
presenting these actions to the user.