Revert "Extract an un-finished autocomplete directive"
This reverts commit 7e4da2ef6d2730a13d5871cbc8437820f8b320d2.

While attempting to clean up the code, this change seems to interfere
with preventing the keyboard from hiding itself.

To demonstrate this:

  1. Tap in the 'Add Item' input field (the keyboard will appear)
  2. Type in an item, then press the '+' button on the UI

Before this reverted commit, the keyboard would stay visible, but this
commit makes it hide when the input field looses focus.

The same problem occurs when selecting an auto-complete suggestion:

  1. Tap in the 'Add Item' input field (the keyboard will appear)
  2. Start typing in an item, tap an auto-suggested item on the UI

The keyboard should stay focused here too.