Speed boost! Use AppCache for offline access
Use the minified versions of the libraries!
This also uses a manual, hand-crafted appcache manifest with all the
files that I could see were actually being used from the Chrome
DevTools.

Hopefully there isn't too much different between browsers, as I've done
no testing with other browsers yet :P

AppCache also needs to have the query params as part of the names, so in
the case of the fonts, I've hardcoded ?v=2.0.1 ... not sure if there is
a better way, but this seems to work for now